upd: renamed layout components to clear the struxure
