+++ src/components/layout/LayoutPortfolio/LayoutPortfolio.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react'
-import { CentralModuleApp } from '@/components/layout/CentralModuleApp/CentralModuleApp'
-import { SidebarApp } from '@/components/layout/SidebarApp/SidebarApp'
+import { CentralModuleApp } from '@/components/layout/LayoutPortfolio/CentralModule/CentralModule'
+import { SidebarApp } from '@/components/layout/LayoutPortfolio/Sidebar/Sidebar'
 
 export function LayoutApp({ children }: PropsWithChildren<unknown>) {
     return (
